Added MapWidgetContainer that contains SearchButton and RecentChipsContainer
